VectorProvider: Split VectorProvider/VectorPipeline classes
